+++ src/lib/ui/CommunityNavItem.svelte
